requirements.txt
docker-py==1.5.0
scp==0.10.2
pyyaml==3.10
GitPython==1.0.1
requests==2.8.1
pyslack==0.1.0
jsonrpc2-zeromq==1.1.2
